3 login logout

+++ app/Http/Controllers/IdeaController.php
@@ -9,9 +9,11 @@ class IdeaController extends Controller
 {
     public function store()
     {
-        request()->validate([ 'content' => 'required|min:5|max:255', ]);
+        $validated = request()->validate(['content' => 'required|min:5|max:255',]);
 
-        $idea = Idea::create([ 'content' => request()->get('content',''), ]);
+        $validated['user_id'] = auth()->id();
+
+        Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea created successfully!');
     }
@@ -23,15 +25,19 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) { // If the authenticated user is not the owner of the idea
+            abort(403);
+        }
+
         $editing = true;
         return view('ideas.show',compact('idea','editing'));
     }
 
     public function update(Idea $idea)
     {
-        request()->validate([ 'content' => 'required|min:5|max:255', ]);
+        $validated = request()->validate([ 'content' => 'required|min:5|max:255', ]);
 
-        $idea->update([ 'content' => request()->get('content',''), ]);
+        $idea->update($validated);
         $idea->save();
 
         return redirect()->route('ideas.show', $idea->id)->with('success', 'Idea updated successfully!');
@@ -39,6 +45,11 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) { // If the authenticated user is not the owner of the idea
+            abort(403);
+//            return back()->with('error', 'You are not allowed to delete this idea!');
+        }
+
         $idea->delete();
         return redirect()->route('dashboard')->with('success', 'Idea deleted successfully!');
     }
